Bug Fix - Solid Element Coord System Default Issue
Bug found and resolved by Victor.
Not sure how this affects thermal expansion (ALPVEC) and anisotropic materials. Also, there might be a cleaner way to write it by calling MATMULT_FFF and MATMULT_FFF_T differently.

The bug can also be worked around in the case of basic coordinates by specifying blank instead of 0 for CID in the PSOLID card. But this workaround doesn't work if you need a different coordinate system.